(Partially) refactor Nix shell configuration for new Codename setup
Codename has upgraded to Haxe 4.3.7 as the minimum recommended version,
and the repository has moved some files around. The Nix shell
configuration has been updated as a result.

Do note that this is only a partial refactor, due to changes on the
UNIX setup file being awaited, as a result of a bug. Alongside this,
although the game compiles normally using this commit's configuration,
the Nix shell configuration uses an outdated Haxe version (4.3.6),
due to 4.3.7 not being available in Nixpkgs at the moment.

An overlay is commented out of the configuration; the reason for this
is that although the download is fine, Nix will attempt to compile the
derivation by default, and Haxe does not appear to compile properly
using that overlay.